+++ api/admin/admin.routes.ts
@@ -1,10 +1,9 @@
 import * as express from "express";
 import { getAllData } from "./admin.controller";
-import { requireAuth } from "../middleware/authMiddleware";
 
 const router = express.Router();
 
-router.get("/", requireAuth, getAllData);
+router.get("/", getAllData);
 
 // router.post("/",  addChef);
 // router.put("/:id",  updateChef);
