+++ api/auth/auth.routes.ts
@@ -1,10 +1,10 @@
 import * as express from "express";
-import { signup, login } from "./auth.controller";
+import { signup, login, logout } from "./auth.controller";
 
 const router = express.Router();
 
 router.post("/signup", signup);
 router.post("/login", login);
-// router.get("logout", logout);
+router.get("/logout", logout);
 
 export default router;
