--- api/middleware/authMiddleware.ts
@@ -1,33 +1,0 @@
-import * as jwt from "jsonwebtoken";
-import { Request, Response, NextFunction } from "express";
-
-const SECRET_KEY = process.env.SECRET_KEY!;
-
-interface DecodedToken {
-    id: string;
-    iat: number;
-    exp: number;
-}
-
-export const requireAuth = (
-    req: Request,
-    res: Response,
-    next: NextFunction
-) => {
-    const token = req.cookies.jwt;
-
-    // check json web token exists & is verified
-    if (token) {
-        jwt.verify(token, SECRET_KEY, (err: any, decodedToken: any) => {
-            if (err) {
-                console.log(err.message);
-                res.redirect("/login");
-            } else {
-                console.log(decodedToken);
-                next();
-            }
-        });
-    } else {
-        res.redirect("/login");
-    }
-};
