+++ logs/backend.log
@@ -1009,3 +1009,62 @@
 15.8.2023, 13:40:11 - DEBUG - Found 12 restaurants for category  
 15.8.2023, 13:50:13 - DEBUG - Found 12 restaurants for category  
 15.8.2023, 14:02:04 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:08:00 - INFO - Server is running on port: 4000 
+15.8.2023, 16:08:06 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:08:57 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:09:27 - INFO - Server is running on port: 4000 
+15.8.2023, 16:09:38 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:09:38 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:09:39 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:09:43 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:10:21 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:10:30 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:10:44 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:11:26 - INFO - Server is running on port: 4000 
+15.8.2023, 16:11:30 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:11:34 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:12:27 - INFO - Server is running on port: 4000 
+15.8.2023, 16:12:30 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:12:49 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:13:21 - INFO - Server is running on port: 4000 
+15.8.2023, 16:16:12 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:16:19 - INFO - Server is running on port: 4000 
+15.8.2023, 16:16:26 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:16:57 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:17:19 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:18:19 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:18:26 - INFO - Server is running on port: 4000 
+15.8.2023, 16:18:29 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:18:30 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:18:32 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:18:33 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:19:04 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:19:06 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:20:06 - INFO - Server is running on port: 4000 
+15.8.2023, 16:20:08 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:20:09 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:21:59 - INFO - Server is running on port: 4000 
+15.8.2023, 16:22:12 - INFO - Server is running on port: 4000 
+15.8.2023, 16:22:45 - INFO - Server is running on port: 4000 
+15.8.2023, 16:23:01 - INFO - Server is running on port: 4000 
+15.8.2023, 16:23:26 - INFO - Server is running on port: 4000 
+15.8.2023, 16:42:00 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:45:08 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:45:36 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:45:42 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:48:58 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:49:07 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:49:25 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:49:32 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:50:10 - INFO - Server is running on port: 4000 
+15.8.2023, 16:52:32 - INFO - Server is running on port: 4000 
+15.8.2023, 16:53:47 - INFO - Server is running on port: 4000 
+15.8.2023, 16:54:42 - INFO - Server is running on port: 4000 
+15.8.2023, 16:55:28 - INFO - Server is running on port: 4000 
+15.8.2023, 16:55:32 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:56:11 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:56:15 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:56:22 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:56:30 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 16:56:56 - DEBUG - Found 12 restaurants for category  
+15.8.2023, 17:00:01 - DEBUG - Found 12 restaurants for category
