+++ routes/index.ts
@@ -1,5 +1,5 @@
 import express, { Express } from "express";
-
+import { checkUser, requireAuth } from "../middleware/authMiddleware";
 import restaurantRoutes from "../api/restaurant/restaurant.routes";
 import dishRoutes from "../api/dish/dish.routes";
 import chefRoutes from "../api/chef/chef.routes";
@@ -8,10 +8,11 @@ import adminRoutes from "../api/admin/admin.routes";
 
 const router = express.Router();
 
+router.get("*", checkUser);
 router.use("/auth", authRoutes);
 router.use("/restaurant", restaurantRoutes);
 router.use("/chef", chefRoutes);
 router.use("/dish", dishRoutes);
-router.use("/admin", adminRoutes);
+router.use("/admin", requireAuth, adminRoutes);
 
 export default router;
